Refragmanting React
